backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
faiss-cpu
sentence-transformers
transformers
huggingface-hub
python-dotenv
python-multipart
pdfplumber
